requirements.txt
appdirs==1.4.3
certifi==2017.4.17
chardet==3.0.4
Django==1.9.13
django-allauth>=0.32.0
django-betterforms>=1.1.4 # TODO change to crispy forms
django-braces>=1.11.0
django-debug-toolbar>=1.8
django-environ>=0.4.3
django-haystack>=2.6.1 # TODO using haystack & whoosh
idna==2.5
oauthlib==2.0.2
packaging==16.8
pyparsing==2.2.0
python-openid==2.2.5
pytz==2017.2
requests==2.18.1
requests-oauthlib==0.8.0
six==1.10.0
sqlparse==0.2.3
urllib3==1.21.1
# TODO include raven(https://docs.sentry.io/clients/python/)
# TODO postgresSQL